+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [0.4.0] - 2020-04-01
+## [0.4.0] - TODO
 ### Added
 - Setting `tracking_roles` added to be used as whitelist for roles.
+- Setting `log_scope` added to decide what should be logged
+### Changed
+- UI of setting `course_ids` changed to `textarea` for easier input
 
 ## [0.3.0] - 2020-03-31
 ### Changed